refactor: split tools/session_manager.py into brokers/ modules (#69)
Extract session management into three focused modules under brokers/:
- brokers/session_pickle.py — SessionPickleManager for pickle read/write/clear/encrypt/decrypt
- brokers/session_mfa.py — MFA resolution, login-prompt handling, device verification
- brokers/session_state.py — SessionManager, module-level functions, singleton

tools/session_manager.py is now a thin re-export shim. All 8 source file
imports and all test patch targets updated to point at the new locations.
