+++ src/open_stocks_mcp/server/http_transport.py
@@ -16,6 +16,7 @@ from mcp.server.fastmcp import FastMCP
 from starlette.middleware.base import BaseHTTPMiddleware
 
 from open_stocks_mcp import __version__
+from open_stocks_mcp.brokers.session_state import get_session_manager
 from open_stocks_mcp.config import load_config
 from open_stocks_mcp.health import get_health_service
 from open_stocks_mcp.logging_config import logger
@@ -27,7 +28,6 @@ from open_stocks_mcp.server.tool_docs import (
 )
 from open_stocks_mcp.tools.circuit_breaker import get_broker_circuit_breaker
 from open_stocks_mcp.tools.rate_limiter import get_rate_limiter
-from open_stocks_mcp.tools.session_manager import get_session_manager
 
 MAX_MCP_REQUEST_BODY_SIZE = 1024 * 1024  # 1 MiB
